tests pass, fuck it
this will go a bit more wild than the tupletable shipped in the distro,
such that foo-amd64 will asume base-gnu-foo-amd64. there's a lot of shit
out there that's pretty wild in the archive, and this appears to be the
intent. This may backfire for things like "windows-foo" if that ever
happens, since that will be "base-gnu-windows-foo" rather than the more
likely "base-nt-window-foo" or what have you.