add in proper source name support
for binnmus or other cases where the binary version isn't the same as
the source version, control files will change from

Source: foo

to

Source: foo (1.0)

where 1.0 is the version of the source package this is generated from.
This is done in cases where there is no source change, but there was a
version change to trigger a rebuild.